chore: split go files.
